pyproject.toml
[project]
name = "gundog"
version = "0.1.0"
description = "Semantic embedding validation tool for ADRs and code"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.12"
authors = [{ name = "ivdi" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "typer>=0.12.0",
    "rich>=13.0",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
lance = ["lancedb>=0.1.0"]
viz = ["pyvis>=0.3.0"]
serve = ["fastapi>=0.115.0", "uvicorn>=0.32.0"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-bdd>=7.0",
    "ruff>=0.8",
    "pyright>=1.1",
    "pre-commit>=3.0",
    "tox>=4.0",
    "tox-uv>=1.0",
    "python-semantic-release>=9.0",
]

[project.scripts]
gundog = "gundog._cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gundog"]

[tool.ruff]
line-length = 99
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "D",       # no docstrings in tests
    "S101",    # allow assert in tests
]

[tool.ruff.format]
quote-style = "double"
line-ending = "lf"

[tool.pyright]
pythonVersion = "3.12"
pythonPlatform = "Linux"
typeCheckingMode = "standard"
include = ["src"]
exclude = ["**/__pycache__", ".venv", "venv", ".tox"]
reportMissingImports = true
reportUnusedImport = true
reportUnusedVariable = true
reportDuplicateImport = true
reportPrivateUsage = true
reportUnnecessaryTypeIgnoreComment = true
reportMissingTypeStubs = false
reportUntypedFunctionDecorator = false

[tool.coverage.run]
source = ["src/gundog"]
branch = true
omit = ["tests/*", "**/__pycache__/*", ".venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-v --tb=short"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variables = ["src/gundog/_version.py:__version__"]
branch = "main"
build_command = "uv build"
commit_message = "chore(release): {version}"
tag_format = "v{version}"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["feat", "fix", "docs", "style", "refactor", "perf", "test", "build", "ci", "chore"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf", "refactor"]

[tool.semantic_release.changelog]
exclude_commit_patterns = ["^chore\\(release\\):"]

[tool.semantic_release.remote]
type = "github"
